Refactor imports for UpsellDevInstances across embedding settings components
