+++ frontend/src/metabase/admin/settings/components/EmbeddingSettings/StaticEmbeddingSettings.tsx
@@ -4,7 +4,7 @@ import {
   SettingsPageWrapper,
   SettingsSection,
 } from "metabase/admin/components/SettingsSection";
-import { UpsellDevInstances } from "metabase/admin/upsells/UpsellDevInstances";
+import { UpsellDevInstances } from "metabase/admin/upsells";
 import { useSetting } from "metabase/common/hooks";
 import { Box } from "metabase/ui";
